add a use demo

+++ test.c
@@ -10,7 +10,7 @@ int main ()
 	struct mq_attr attr;
 	char filename [] = "testfile";
 	char receive[1024];
-	int proi;
+	unsigned int proi;
 	ssize_t n;
 	mqd_t mqd;
 
@@ -24,9 +24,9 @@ int main ()
 	printf ("%ld %ld %ld\n",attr.mq_maxmsg,attr.mq_msgsize,attr.mq_curmsgs);
 
 	mq_send (mqd,"1234",4,10);
-	n = mq_receive (mqd,receive,1024,&proi);
+	n = mq_receive (mqd,receive,(size_t)1024,&proi);
 
-	printf ("%d %s\n",n,receive);
+	printf ("%ld %s\n",n,receive);
 	mq_close (mqd);
 	mq_unlink (filename);
